fix: add GoalAdder component and refactor related code
Introduced a new GoalAdder component to improve the goal addition process.
Refactored the related state and functions to integrate with the new component.
Fixed CSS flex positioning issues, including gap and alignment adjustments.
Refactored the IconButton component for improved reusability across the project.
