+++ frontend/lmapp/src/components/Goals/GoalsList.js
@@ -1,25 +1,19 @@
-import React, { useState, useMemo } from "react";
-import MyIcon from "../Icons/MyIcon";
-import AddIcon from '@mui/icons-material/Add';
+import { useState, useMemo } from "react";
 
 import List from '../List/List';
 import TabButton from './TabButton';
+import GoalAdder from "./GoalAdder";
 import useList from '../../hooks/useList';
 import '../../styles/Goals/goalsList.css';
+
 
 const GoalsList = () => {
 
   const { mockList } = useList();
   const [goals, setGoals] = useState(mockList);
   const [activeTab, setActiveTab] = useState('All');
-  const [text, setText] = useState(''); //Update the text for the new goal
 
-  const handleChange = (e) => {
-    //console.log(e.target.value);
-    setText(e.target.value);
-  }
-
-  const addGoal = () => {
+  const addGoal = (text) => {
     // create a new goal object
     const newGoal = {
       id: goals.length + 1,
@@ -30,13 +24,14 @@ const GoalsList = () => {
     if (text !== "") {
       // add new goal to the list (array)
       setGoals([...goals, newGoal]);
-      setText('');
       //console.log('goal added');
     }
   }
 
-  const deleteGoal = (id) => {
-    setGoals(goals.filter(goal => goal.id !== id));
+  const editGoal = (id, itemText) => {
+    setGoals(goals.map(goal => {
+      return goal.id === id ? { ...goal, text: itemText } : { ...goal };
+    }));
   }
 
   const updateGoal = (id) => {
@@ -45,10 +40,8 @@ const GoalsList = () => {
     }));
   }
 
-  const editGoal = (id, itemText) => {
-    setGoals(goals.map(goal => {
-      return goal.id === id ? { ...goal, text: itemText } : { ...goal };
-    }));
+  const deleteGoal = (id) => {
+    setGoals(goals.filter(goal => goal.id !== id));
   }
 
   const filteredGoals = useMemo(() => {
@@ -71,16 +64,7 @@ const GoalsList = () => {
           <TabButton text="Pending" onClick={() => setActiveTab('Pending')} active={activeTab === 'Pending'} />
           <TabButton text="Completed" onClick={() => setActiveTab('Completed')} active={activeTab === 'Completed'} />
         </div>
-        <input
-          type="text"
-          className="input-login"
-          placeholder="input text for the new goal"
-          name="add-goal-textbox"
-          onChange={handleChange}
-        />
-        <div className="plus-button">
-          <MyIcon icon={AddIcon} size="20px" backgroundColor='lightgreen' color='white' onClick={addGoal} />
-        </div>
+        <GoalAdder addGoal={addGoal} />
       </div>
       <div className="goals-list-container">
         <div className="goals-container">
